+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 [Semantic Versioning](https://semver.org/)
 
-## [Unreleased]
+## [1.39.6] - 2026-06-30
 
 ### Fix